Move ORM models to `db/models`
