+++ backend/app/api/routers/machine.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
 from app.api.deps import get_db, transaction
-from app.db.machine import Machine
+from app.db.models.machine import Machine
 from app.schemas import MachineCreate, MachineOut
 
 router = APIRouter(prefix="/machines", tags=["Machines"])
