+++ backend/app/db/models/artifact.py
@@ -8,11 +8,11 @@ from sqlalchemy import ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from app.db.base import Base
 from app.db.mixins import IDMixin, TimestampMixin
+from app.db.models.base import Base
 
 if TYPE_CHECKING:
-    from app.db.simulation import Simulation
+    from app.db.models.simulation import Simulation
 
 
 class ArtifactKind(str, Enum):
