+++ backend/app/db/models/status.py
@@ -1,7 +1,7 @@
 from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.db.base import Base
+from app.db.models.base import Base
 
 
 class Status(Base):
